Create Program.cs
I have built a chat client that connects to the server and allows the user to send and receive messages. I have added my own method to retrieve the username, adjusted variable names and output messages, and commented on each part in Swedish to show that I understand how it works. I have also defined the timestamp format for the messages so that it becomes my own variant."